+++ src/components/DeviceList.vue
@@ -1,24 +1,32 @@
 <template>
   <q-list separator>
-    <q-list-header>
+    <q-window-resize-observable @resize="onResize" />
+    <q-list-header ref="header">
       <big>
         <q-icon name="mdi-arrow-left" size="1.8rem" class="cursor-pointer" style="margin-right: 15px" @click.native="$emit('click:hide')"/>
         Devices
       </big>
     </q-list-header>
-    <device @update:watch-by-id="setWatchByDeviceID" v-for="device in devices" :key="device.id" :device="device" :activeDevicesID="activeDevicesID" :isDeviceWatched="deviceIdForWatch === device.id"></device>
+    <VirtualList
+      :size="70"
+      :remain="itemsCount"
+    >
+      <device @update:watch-by-id="setWatchByDeviceID" v-for="device in devices" :key="device.id" :device="device" :activeDevicesID="activeDevicesID" :isDeviceWatched="deviceIdForWatch === device.id" />
+    </VirtualList>
   </q-list>
 </template>
 
 <script>
 import { mapActions } from 'vuex'
 import Device from './Device.vue'
+import VirtualList from 'vue-virtual-scroll-list'
 
 export default {
   name: 'DeviceList',
   data () {
     return {
-      unsubscribeDevices: () => {}
+      unsubscribeDevices: () => {},
+      itemsCount: 0
     }
   },
   props: [
@@ -27,7 +35,8 @@ export default {
     'deviceIdForWatch'
   ],
   components: {
-    Device
+    Device,
+    VirtualList
   },
   methods: {
     ...mapActions([
@@ -35,6 +44,9 @@ export default {
     ]),
     setWatchByDeviceID (id) {
       this.$emit('update:watch-by-id', id)
+    },
+    onResize ({height}) {
+      this.itemsCount = (height - 76) / 70
     }
   },
   async created () {